simplify is_categorical test
